Kernel: pcibarsize
While working on this, it was discovered that the pci
code assimed the smallest I/O bar size is 16 (pcibarsize()),
which is wrong. I/O bars can be as small as 4 bytes.
Bit 1 in an I/O bar is also reserved and should be masked off,
making the port mask: port = bar & ~3;

https://github.com/9front/9front/commit/4f85115526a87063489dc7cf347343bd520159b1